+++ Cargo.toml
@@ -12,7 +12,6 @@ repository = "https://github.com/flipt-io/ffs"
 [dependencies]
 anyhow = "1.0.70"
 clap = { version = "4.2.0", features = ["derive"] }
-flipt = "0.4.0"
 human-panic = "1.1.3"
 serde = "1.0.159"
 serde_json = "1.0.95"
